[초기세팅] 화면 사이즈 layout 분리  (#28)
* feat: 대쉬보드 레이아웃 변경 및 사이즈 스타일 적용

* feat: app 리사이즈 useEffect 제거

* feat: useEffect 제거 후 스타일 고정

* feat: 레이아웃 커스텀 훅 구현

* feat: 라우터 폴더 구조 변경

* feat: layout 폴더 구조 변경
