--- src/components/common/DashBoard.tsx
@@ -1,13 +1,0 @@
-import { Outlet } from "react-router-dom";
-import Footer from "./Footer";
-import Header from "./Header";
-
-export default function DashBoard() {
-  return (
-    <>
-      <Header />
-      <Outlet />
-      <Footer />
-    </>
-  );
-}
