+++ src/router/Router.tsx
@@ -1,0 +1,23 @@
+import { createBrowserRouter } from 'react-router-dom';
+import Main from '../pages/Main';
+import Mypage from '../pages/Mypage';
+import Layout from '../layouts/Layout';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Layout />,
+    children: [
+      {
+        path: '/',
+        element: <Main />,
+      },
+      {
+        path: '/mypage',
+        element: <Mypage />,
+      },
+    ],
+  },
+]);
+
+export default router;
